Remove unused method and rename AlertSteps
